ocr-server/requirements.txt
flask==3.0.3
easyocr==1.7.2
Pillow==10.4.0
numpy==1.26.4
